requirements.txt
fastapi
uvicorn
pytest
pytest-cov
python-multipart
pytest-dotenv
httpx
sqlalchemy
pika
psycopg2-binary
pydantic[email]
python-dotenv
azure-storage-blob==12.16.0